requirements.txt
asgiref==3.7.2
bleach==6.1.0
boto3==1.33.5
botocore==1.33.5
dj-database-url==2.1.0
Django==4.2.6
django-cors-headers==4.3.0
django-environ==0.11.2
django-storages==1.14.2
django-subdomains==2.1.0
django-summernote==0.8.20.0
django-tenant-schemas==1.12.0
django-tenants==3.5.0
django-tinymce==3.6.1
djangorestframework==3.14.0
jmespath==1.0.1
ordered-set==4.1.0
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
s3transfer==0.8.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
webencodings==0.5.1
whitenoise==6.6.0
